chore(chats): splitted delete chat components
