+++ src/components/sidebar/components/sidebarDesktop/SidebarDesktopDeleteChat/types.ts
@@ -1,0 +1,8 @@
+export interface DeleteChatProps {
+  chatId: string
+  isLastChat: boolean
+}
+
+export interface SidebarDesktopDeleteChatProps extends DeleteChatProps {
+  isPrivate?: boolean
+}
